requirements.txt
# Core Libraries
requests==2.28.2
pandas==2.0.1
selenium==4.9.0
webdriver-manager==3.8.6

# CLI and Output
rich==13.3.5
click==8.1.3
tqdm==4.65.0

# Environment and Configuration
python-dotenv==1.0.0

# Development and Testing
pytest==7.3.1